Update README.md
A total of six countries were analyzed for the effect and extent at which the natural disasters affect them. they are China, Brazil, USA, Indonesia, India and Japan. it was observed however, that the earhquake occured the most with about 20.4% occurence rate closely followed by flood with a 20.2%, Tornado with a 20.1%, wildfire 19.8% and a Hurricane with 19.5% occurence. The fatalities recorded were over 49million cases with the fatality due to Earthquake topping the list at 1095826 cases and closely followed by the cases of flood; 10086080 cases, Tornado recorded 10076567, hurricance; 9883502 and wildfire the least with 9707314 cases in total. 

Although the earthquake had the most occurence as recorded above, the effect of the flooding were worse o the economy when compared with the other disasters. Also Brazil, was the mose hit and affected economically as they had the most frequency of natural disasters with a total of 1743 times out of the total of 10000 occurences. 

However to curtail the occurences of these disasters and mitigate the effect on the country, its economy and people at large, the following recommendations were made; 

Priortise high risk disaster awareness and put measures in place to avoid such high levelof occurences and also try to identify patterns in disaster prone areas, building safer structures and providing education on earthquake safety.
Governemnt and organizations can also focus on earthquake resistant constructions and more especially in regions prone to seismic activity. 
Since flooding resulted in the highest economic losses, it should be the focus of early warning systems, preparedness training, and infrastructure reinforcements like dams, leeves, bridges etc where necessary to manage and mitigatethe occurence of flooding.
The goverment of Brazil in conjuntion with concerned bodies are adviced to look out for ways ike insurance and financial relief mechanisms to keep the economy afloat so it does not affect the cicitzens of the country in the long run,